tests/wpt/mozilla/tests/bluetooth/requestDevice/single-filter-single-service.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.glucose_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [glucose.name]}]
    })
    .then(device => assert_equals(device.name, mock_device_name.glucose));
}, 'Simple filter selects matching device.');
</script>
